Move ResolvingValue to DefaultTaskProperties
